update path to edit predios page

+++ apollo/graphql/schemas/predio.ts
@@ -3,7 +3,7 @@ export const Predio = `type Predio {
   lotName: String!
   lotDepartment: String!
   lotMunicipality: String!
-  lotValue: Int!
+  lotValue: Float!
   constructions: [Construccion!]
   owners: [Persona!]
   land: [Terreno!]
@@ -14,5 +14,5 @@ export const PredioInput = `input PredioInput{
   lotName: String
   lotDepartment: String
   lotMunicipality: String
-  lotValue: Int
+  lotValue: Float
 }`
